+++ src/egraphs/unionfind.rs
@@ -11,12 +11,12 @@ use std::collections::HashMap;
 #[derive(Debug, Clone)]
 pub(crate) enum ProofTrackerValue {
     Root { size: i32 },
-    Parent { parent: u64 },
+    Parent { parent: u32 },
 }
 
 /// Forest of the different proof trackers
 pub struct ProofTracker {
-    forest: HashMap<u64, ProofTrackerValue>,
+    forest: HashMap<u32, ProofTrackerValue>,
 }
 
 impl ProofTracker {
@@ -27,7 +27,7 @@ impl ProofTracker {
     }
 
     /// unioning terms x and y
-    pub fn union(&mut self, x: u64, y: u64) -> bool {
+    pub fn union(&mut self, x: u32, y: u32) -> bool {
         debug_println!(
             5,
             2,
@@ -114,7 +114,7 @@ impl ProofTracker {
         true
     }
 
-    fn find(&mut self, x: u64) -> u64 {
+    fn find(&mut self, x: u32) -> u32 {
         debug_println!(
             5,
             0,
